move register page constants to separate file
