+++ travel-journal-ui/src/components/Auth/Register/Register.jsx
@@ -8,23 +8,7 @@ import "./index.css";
 import logo from "../../../assets/Logo-White.svg";
 import login_bg from "../../../assets/login-bg.jpg";
 
-// the keys have the same name as the input states
-const REGEX = {
-    firstName: /^[a-zA-Z]+$/,
-    lastName: /^[a-zA-Z]+$/,
-    email: /(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\.){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])/,
-    password: /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z]).{8,}$/,
-}
-
-const errorMessages = {
-    firstName: "First Name must contain only letters and no spaces",
-    lastName: "Last Name must contain only letters and no spaces",
-    email: "Invalid email format",
-    password: "Password must contain at least 8 characters including at least one upper and lowercase letter and a number.",
-    confirmPassword: "Passwords do not match",
-    emptyField: "This field is required",
-}
-
+import {REGEX, errorMessages} from "../../../constants/Validations.js"
 
 const Register = () => {
     const navigate = useNavigate();
